Back out 1e7050c52cb9 (bug 792850), 776f7325952b (bug 705047), 2e1be0680e70 (bug 733501), 6adef296963c (bug 791775), b8397781f862 (bug 792391) for WinXP hangs and suspicion of breaking Android stacks

+++ toolkit/crashreporter/google-breakpad/src/common/dwarf/bytereader.h
@@ -85,7 +85,7 @@ class ByteReader {
   //
   // - If N is between 0 and 0x7f, then its unsigned LEB128
   //   representation is a single byte whose value is N.
-  //
+  // 
   // - Otherwise, its unsigned LEB128 representation is (N & 0x7f) |
   //   0x80, followed by the unsigned LEB128 representation of N /
   //   128, rounded towards negative infinity.
@@ -104,7 +104,7 @@ class ByteReader {
   // - If N is between -0x40 and 0x3f, then its signed LEB128
   //   representation is a single byte whose value is N in two's
   //   complement.
-  //
+  // 
   // - Otherwise, its signed LEB128 representation is (N & 0x7f) |
   //   0x80, followed by the signed LEB128 representation of N / 128,
   //   rounded towards negative infinity.
@@ -301,7 +301,7 @@ class ByteReader {
   // Base addresses for Linux C++ exception handling data's encoded pointers.
   bool have_section_base_, have_text_base_, have_data_base_;
   bool have_function_base_;
-  uint64 section_base_, text_base_, data_base_, function_base_;
+  size_t section_base_, text_base_, data_base_, function_base_;
   const char *buffer_base_;
 };
